Move tests to proper packages
