+++ app/src/androidTest/java/io/github/wulkanowy/data/repositories/recipient/RecipientLocalTest.kt
@@ -1,4 +1,4 @@
-package io.github.wulkanowy.data.repositories.local
+package io.github.wulkanowy.data.repositories.recipient
 
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
@@ -7,7 +7,6 @@ import io.github.wulkanowy.data.db.AppDatabase
 import io.github.wulkanowy.data.db.entities.Recipient
 import io.github.wulkanowy.data.db.entities.ReportingUnit
 import io.github.wulkanowy.data.db.entities.Student
-import io.github.wulkanowy.data.repositories.recipient.RecipientLocal
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
